+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/valibot
 
+## 3.0.0-canary.32
+
+### Patch Changes
+
+- Updated dependencies [5463d0d]
+  - @ai-sdk/provider-utils@5.0.0-canary.32
+
 ## 3.0.0-canary.31
 
 ### Patch Changes